requirements.txt
cppimport==22.8.2
networkx==2.8.8
numpy==1.24.3
prettytable==3.7.0
scikit_learn==1.2.2
scipy==1.10.1
setproctitle==1.3.2
torch_geometric==2.1.0
torch_scatter==2.1.0
tqdm==4.65.0
